moved datatypes
