+++ client-src/elements/autolink.ts
@@ -3,6 +3,7 @@
 // See: https://chromium.googlesource.com/infra/infra/+/refs/heads/main/appengine/monorail/static_src/autolink.js
 
 import {enhanceAutolink} from './chromedash-link.js';
+import {Component, TextRun} from './datatypes.ts';
 const CRBUG_DEFAULT_PROJECT = 'chromium';
 const CRBUG_URL = 'https://bugs.chromium.org';
 const ISSUE_TRACKER_RE =
@@ -213,15 +214,4 @@ function applyLinks(textRuns, replacer, re) {
     }
   });
   return resultRuns;
-}
-
-interface TextRun {
-  content: string;
-  tag?: string;
-  href?: string;
-}
-
-interface Component {
-  refRegs: RegExp[];
-  replacer: (match: any) => never[] | TextRun[];
 }
